Implemented package-info.java files for most major packages and did some organizational moves for classes that could be in a more logical place.
