+++ src/main/java/org/vizun/Vizun.java
@@ -3,7 +3,6 @@ package org.vizun;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.vizun.engine.Game;
-import org.vizun.engine.events.GameHandler;
 import org.lwjgl.opengl.Display;
 import org.vizun.lib.DataFolder;
 
@@ -13,6 +12,8 @@ public class Vizun implements Game {
     private static final GameHandler gameHandler = new GameHandler(instance);
     
     private static DataFolder dataFolder;
+    
+    
 
     public static void main(String[] args){
         final Vizun vizun = new Vizun();
@@ -49,6 +50,4 @@ public class Vizun implements Game {
     public static DataFolder getDataFolder() {
         return dataFolder;
     }
-
-
 }
