+++ src/main/java/org/vizun/exception/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Custom client and server exceptions used by the Vizun Project
+ */
+package org.vizun.exception;
