+++ src/main/java/org/vizun/lib/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Libraries and tools that require instantiation and cannot be used in a static manner
+ */
+package org.vizun.lib;
