+++ src/main/java/org/vizun/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Main Package for the Vizun Game
+ */
+package org.vizun;
