+++ src/main/java/org/vizun/util/LogSend.java
@@ -1,0 +1,7 @@
+package org.vizun.util;
+
+/**
+ * This class will handle the sending of log files to the remote master server for *
+ */
+public class LogSend {
+}
