+++ src/main/java/org/vizun/util/Math.java
@@ -1,4 +1,4 @@
-package org.vizun.toolbox;
+package org.vizun.util;
 
 /**
  * Created by River on 5/17/2015.
