+++ src/main/java/org/vizun/util/VoxelData.java
@@ -1,4 +1,4 @@
-package org.vizun.engine.util;
+package org.vizun.util;
 
 /**
  * Created by James on 5/16/2015.
