+++ src/main/java/org/vizun/util/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Utilities and Static Libraries used for a specific task.
+ */
+package org.vizun.util;
